@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.2</TargetFramework>
+        <TargetFramework>netcoreapp3.0</TargetFramework>
         <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
         <TypeScriptToolsVersion>Latest</TypeScriptToolsVersion>
         <TypeScriptCompileBlocked>true</TypeScriptCompileBlocked>
@@ -19,8 +19,6 @@
     <ItemGroup>
         <PackageReference Include="Dapper" Version="2.0.30" />
         <PackageReference Include="Markdig" Version="0.17.1" />
-        <PackageReference Include="Microsoft.AspNetCore.App" />
-        <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
         <PackageReference Include="Microsoft.Data.SqlClient" Version="1.0.19269.1" />
         <PackageReference Include="Microsoft.DocAsCode.MarkdownLite" Version="2.45.1" />
         <PackageReference Include="StyleCop.Analyzers" Version="1.1.118">
